Revert "Switch to BigDecimal for more precision"
This reverts commit 8d0b63b1ab44afaad321e187ad89555990786d1b.

+++ src/main/java/be/robbevanherck/javafraggenescan/entities/HMMParameters.java
@@ -3,26 +3,25 @@ package be.robbevanherck.javafraggenescan.entities;
 import be.robbevanherck.javafraggenescan.repositories.*;
 
 import java.io.File;
-import java.math.BigDecimal;
 import java.util.Map;
 
 /**
  * Contains all the parameters for the Hidden Markov Model
  */
 public class HMMParameters {
-    private final Map<HMMInnerTransition, BigDecimal> innerTransitions;
-    private final Map<HMMOuterTransition, BigDecimal> outerTransitions;
-    private final Map<HMMState, Map<Triple<AminoAcid>, BigDecimal>> forwardMatchEmissions;
-    private final Map<HMMState, Map<Triple<AminoAcid>, BigDecimal>> reverseMatchEmissions;
-    private final Map<Pair<AminoAcid>, BigDecimal> insertInsertEmissions;
-    private final Map<Pair<AminoAcid>, BigDecimal> matchInsertEmissions;
-    private final Map<Pair<AminoAcid>, BigDecimal> nonCodingNonCodingEmissions;
+    private final Map<HMMInnerTransition, Double> innerTransitions;
+    private final Map<HMMOuterTransition, Double> outerTransitions;
+    private final Map<HMMState, Map<Triple<AminoAcid>, Double>> forwardMatchEmissions;
+    private final Map<HMMState, Map<Triple<AminoAcid>, Double>> reverseMatchEmissions;
+    private final Map<Pair<AminoAcid>, Double> insertInsertEmissions;
+    private final Map<Pair<AminoAcid>, Double> matchInsertEmissions;
+    private final Map<Pair<AminoAcid>, Double> nonCodingNonCodingEmissions;
     private final boolean wholeGenome;
-    private final Map<Integer, Map<Triple<AminoAcid>, BigDecimal>> forwardStopPWM;
-    private final Map<Integer, Map<Triple<AminoAcid>, BigDecimal>> reverseStopPWM;
-    private final Map<Integer, Map<Triple<AminoAcid>, BigDecimal>> forwardStartPWM;
-    private final Map<Integer, Map<Triple<AminoAcid>, BigDecimal>> reverseStartPWM;
-    private final Map<HMMState, Map<GaussianArgumentsRepository.GaussianArgument, BigDecimal>> gaussianArguments;
+    private final Map<Integer, Map<Triple<AminoAcid>, Double>> forwardStopPWM;
+    private final Map<Integer, Map<Triple<AminoAcid>, Double>> reverseStopPWM;
+    private final Map<Integer, Map<Triple<AminoAcid>, Double>> forwardStartPWM;
+    private final Map<Integer, Map<Triple<AminoAcid>, Double>> reverseStartPWM;
+    private final Map<HMMState, Map<GaussianArgumentsRepository.GaussianArgument, Double>> gaussianArguments;
 
     /**
      * Protected constructor only for testing
@@ -106,7 +105,7 @@ public class HMMParameters {
      * @param state The state the transition is going to
      * @return The probability
      */
-    public BigDecimal getMatchEmissionProbability(HMMState state, Triple<AminoAcid> aminoAcidEndingInT) {
+    public double getMatchEmissionProbability(HMMState state, Triple<AminoAcid> aminoAcidEndingInT) {
         return this.forwardMatchEmissions.get(state).get(aminoAcidEndingInT);
     }
 
@@ -115,7 +114,7 @@ public class HMMParameters {
      * @param transition The transition
      * @return The probability for that transition
      */
-    public BigDecimal getInnerTransitionProbability(HMMInnerTransition transition) {
+    public double getInnerTransitionProbability(HMMInnerTransition transition) {
         return innerTransitions.get(transition);
     }
 
@@ -124,7 +123,7 @@ public class HMMParameters {
      * @param transition The transition
      * @return The probability for that transition
      */
-    public BigDecimal getOuterTransitionProbability(HMMOuterTransition transition) {
+    public double getOuterTransitionProbability(HMMOuterTransition transition) {
         return outerTransitions.get(transition);
     }
 
@@ -134,7 +133,7 @@ public class HMMParameters {
      * @param currentInput The input of the current state
      * @return The probability
      */
-    public BigDecimal getInsertInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
+    public double getInsertInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
         return insertInsertEmissions.get(new Pair<>(previousInput, currentInput));
     }
 
@@ -144,7 +143,7 @@ public class HMMParameters {
      * @param currentInput The input of the current state
      * @return The probability
      */
-    public BigDecimal getMatchInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
+    public double getMatchInsertEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
         return matchInsertEmissions.get(new Pair<>(previousInput, currentInput));
     }
 
@@ -155,7 +154,7 @@ public class HMMParameters {
      * @param currentInput The input of the current state
      * @return The probability
      */
-    public BigDecimal getNonCodingNonCodingEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
+    public double getNonCodingNonCodingEmissionProbability(AminoAcid previousInput, AminoAcid currentInput) {
         return nonCodingNonCodingEmissions.get(new Pair<>(previousInput, currentInput));
     }
 
@@ -166,7 +165,7 @@ public class HMMParameters {
      * @param state The state the transition is going to
      * @return The probability
      */
-    public BigDecimal getReverseMatchEmissionProbability(HMMState state, Triple<AminoAcid> aminoAcidEndingInT) {
+    public double getReverseMatchEmissionProbability(HMMState state, Triple<AminoAcid> aminoAcidEndingInT) {
         return reverseMatchEmissions.get(state).get(aminoAcidEndingInT);
     }
 
@@ -176,7 +175,7 @@ public class HMMParameters {
      * @param codon The codon at the position index
      * @return The probability
      */
-    public BigDecimal getForwardEndPWMProbability(int index, Triple<AminoAcid> codon) {
+    public double getForwardEndPWMProbability(int index, Triple<AminoAcid> codon) {
         return forwardStopPWM.get(index).get(codon);
     }
 
@@ -186,7 +185,7 @@ public class HMMParameters {
      * @param codon The codon at the position index
      * @return The probability
      */
-    public BigDecimal getReverseEndPWMProbability(int index, Triple<AminoAcid> codon) {
+    public double getReverseEndPWMProbability(int index, Triple<AminoAcid> codon) {
         return reverseStopPWM.get(index).get(codon);
     }
 
@@ -196,7 +195,7 @@ public class HMMParameters {
      * @param codon The codon at the position index
      * @return The probability
      */
-    public BigDecimal getForwardStartPWMProbability(int index, Triple<AminoAcid> codon) {
+    public double getForwardStartPWMProbability(int index, Triple<AminoAcid> codon) {
         return forwardStartPWM.get(index).get(codon);
     }
 
@@ -206,7 +205,7 @@ public class HMMParameters {
      * @param codon The codon at the position index
      * @return The probability
      */
-    public BigDecimal getReverseStartPWMProbability(int index, Triple<AminoAcid> codon) {
+    public double getReverseStartPWMProbability(int index, Triple<AminoAcid> codon) {
         return reverseStartPWM.get(index).get(codon);
     }
 
@@ -216,7 +215,7 @@ public class HMMParameters {
      * @param argument The argument to get
      * @return The value
      */
-    public BigDecimal getGaussianArgument(HMMState state, GaussianArgumentsRepository.GaussianArgument argument) {
+    public double getGaussianArgument(HMMState state, GaussianArgumentsRepository.GaussianArgument argument) {
         return gaussianArguments.get(state).get(argument);
     }
 }
