+++ src/test/java/be/robbevanherck/javafraggenescan/dummies/DummyTransition.java
@@ -5,8 +5,6 @@ import be.robbevanherck.javafraggenescan.entities.PathProbability;
 import be.robbevanherck.javafraggenescan.entities.ViterbiStep;
 import be.robbevanherck.javafraggenescan.transitions.Transition;
 
-import java.math.BigDecimal;
-
 public class DummyTransition extends Transition {
     public PathProbability pathProbability;
 
@@ -15,7 +13,7 @@ public class DummyTransition extends Transition {
      */
     public DummyTransition() {
         super(HMMState.NO_STATE);
-        pathProbability = new PathProbability(HMMState.MATCH_1, BigDecimal.valueOf(0.99));
+        pathProbability = new PathProbability(HMMState.MATCH_1, 0.99);
     }
 
     @Override
